basic example working with nextjs pages
