+++ babel.config.js
@@ -18,6 +18,7 @@ module.exports = function (api) {
 			],
 		].filter(Boolean),
 		plugins: [
+			['@babel/plugin-transform-flow-strip-types'],
 			['@babel/plugin-proposal-class-properties', { loose: true }],
 			['@babel/plugin-proposal-private-methods', { loose: true }],
 			['@babel/plugin-proposal-private-property-in-object', { loose: true }],
